requirements.txt
Django>=1.9.4
Django-allauth>=0.25
Pillow>=3.2.0
djangorestframework
markdown
django-filter
django-rest-swagger
twilio
django-rest-auth
